refactor(x64): typed actor/geometry pointers, offset asserts, drag golden (Stage 2)
Six related Stage 2 changes. Pointer width does not change arithmetic, so the
golden traces must stay GREEN throughout -- they are the proof, not just a
regression check.

1. CAMERA API takes a pointer, not an int (td5_camera.h/.c, td5_input.c).
   Ten functions took the actor in an `int` parameter and dereferenced it
   (`*(short *)(actor + 2)`, `*(uint8_t *)(actor + 0x375)`, ...), with ~30 call
   sites truncating a real pointer to 32 bits. Now `uint8_t *`.
   terrain_probe_trace_emit's `int actor_addr` converted likewise.

2. BUGFIX in ResetRaceCameraSelectionState: two LoadCameraPresetForView calls
   passed `(char *)&g_actorBaseAddr + slot*0x388` -- the ADDRESS OF the global
   handle, not the table it names. A wild pointer, harmless only because that
   function dereferences the actor solely when force_reload != 0 and both calls
   pass 0. Now uses g_actor_table_base. Goldens unmoved, which confirms the
   deref never fired.

3. DELETED g_actorBaseAddr (td5_game.c/.h). An `int` holding a truncated
   pointer, redundant with g_actor_table_base / g_actor_base / g_actor_pool set
   beside it. Its last two readers were the camera sites above.

4. g_spanTable / g_vertexTable retyped `int` -> `const uint8_t *`
   (td5_camera.c). These do real address arithmetic (`base + idx*0x18`) and
   then dereference, so an int base is a wild pointer on x64. Locals carrying
   those addresses retyped with them.

5. ACTOR OFFSET CROSS-CHECKS (td5_ai_internal.h). td5_ai.c / td5_ai_traffic.c
   address the actor as a raw byte blob and never included the typed header, so
   their ~25 ACTOR_* constants were UNGUARDED DUPLICATES of the layout in
   re/include/td5_actor_struct.h -- if a field moved they went stale silently.
   Added 26 _Static_asserts tying each to offsetof(TD5_Actor, ...). Zero
   codegen. On x64 the 17 constants at or after +0x1B0 will fail here (the four
   void* grow, shifting everything after by 16 bytes) -- the compiler hands you
   the worklist instead of you debugging wrong AI behaviour.
   Kept the literals rather than redefining them AS offsetof: they double as
   [CONFIRMED @ addr] documentation of the original layout. Self-adapting
   constants are a safe follow-up if preferred; the asserts make it so.

6. SAVE STRIDE ISOLATED (td5_save.c). The hardcoded 0x388 there is an ON-DISK
   stride that merely happens to equal the in-memory one on i686 -- exactly the
   trap a "make stride = sizeof" cleanup would spring. Now the frozen literal
   TD5_ACTOR_SAVE_STRIDE, documented as legacy-binary-only (the live cup format
   is text INI and does not store the actor snapshot). Also corrected the
   actor_state comment: it claimed "6 slots x 0x388", but 12656 bytes is 14
   slots, not 6 (6 x 904 = 5424). Asserted the divisibility only.

VERIFICATION
Changes 1-4 verified: full selftest, both trace-golden sets 7/7 module hashes.
Changes 5-6 are codegen-neutral (dev exe byte-identical in size).
Suite must be run at REDUCED RESOLUTION (1280x676, same aspect) -- the
pre-existing nvwgf2um.dll fault at race-moscow-base is load-dependent.

*** THE DRAG GOLDEN (7) IS RECORDED BUT NOT VERIFIED ***
Added race-golden-drag to k_races (td5_selftest.c) because no golden covered
the drag/gantry path -- and that path builds its display list DIFFERENTLY
(td5_render.c synthesizes a 1-entry block on the stack instead of reading one
from the models blob), so a change to the display-list slot representation can
break drag rendering while both existing goldens stay green. That gap blocks
the mesh-header split.
The recording run was clean (55 PASS / 0 FAIL, 3/3 scenarios) and the trace
diff was additions-only -- all 14 existing sim hashes byte-identical, which is
good evidence the scenario is well-formed and deterministic. But the
independent confirm-green pass CRASHED in nvwgf2um.dll before reaching it, so
race-golden-drag has never been checked against a second run. Re-verify before
relying on it.
Note also render_goldens.txt is now recorded at 1280x676 (was 2560x1351).
Cross-resolution comparison passed comfortably before (worst L1 176/282 vs
limit 4000) so it should still pass at full res, but that is inference, not a
measurement.

RE basis: N/A -- port-side storage representation; original behaviour unchanged

via [Happy](https://happy.engineering)

+++ td5mod/src/td5re/td5_input.c
@@ -942,7 +942,7 @@ void td5_input_poll_race_session(void)
                 }
                 if (actor) {
                     LoadCameraPresetForView(
-                        (int)((uint8_t *)actor + 0x208), 0, cam_view, 1);
+                        (uint8_t *)actor + 0x208, 0, cam_view, 1);
                 }
             }
             s_camera_cooldown[i] = TD5_INPUT_CAMERA_COOLDOWN;
@@ -1171,7 +1171,7 @@ post_poll:
                      * (Same transition rationale as the camera-button branch
                      * above.) */
                     LoadCameraPresetForView(
-                        (int)((uint8_t *)actor + 0x208), 0, i, 1);
+                        (uint8_t *)actor + 0x208, 0, i, 1);
                 }
             }
             s_camera_cooldown[i] = TD5_INPUT_CAMERA_COOLDOWN;
